pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "spoofspy"
dynamic = ["version"]
description = "Experimental app for detecting and displaying spoofed redirect servers on Steam."
readme = "README.md"
requires-python = ">=3.11"
license-files = { paths = ["LICENSE"] }
keywords = []
authors = [
  { name = "Tuomo", email = "tuokri@tuta.io" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

# TODO: split dependencies into optional groups
#  based on sub-packages.
dependencies = [
  "SQLAlchemy[asyncio]==2.0.20",
  "celery[redis]==5.3.1",
  "eventlet==0.33.3",
  "fastapi-cache2==0.2.1",
  "fastapi==0.103.0",
  "httpx==0.24.1",
  "msgpack==1.0.5",
  "numpy==1.25.2",
  "orjson==3.9.5",
  "psycopg[pool]==3.1.10",
  "python-a2s==1.3.0",
  "steam==1.4.4",
  "uvicorn[standard]==0.23.2",
  # "aiohttp[speedups]==3.8.5",
  # "ujson==5.8.0",
]

[project.urls]
Documentation = "https://github.com/tuokri/spoofspy#readme"
Issues = "https://github.com/tuokri/spoofspy/issues"
Source = "https://github.com/tuokri/spoofspy"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "spoofspy/_version.py"

#[tool.hatch.envs.default]
#dependencies = [
#  "coverage[toml]>=6.5",
#  "pytest",
#]
#
#[tool.hatch.envs.default.scripts]
#test = "pytest {args:tests}"
#test-cov = "coverage run -m pytest {args:tests}"
#cov-report = [
#  "- coverage combine",
#  "coverage report",
#]
#cov = [
#  "test-cov",
#  "cov-report",
#]

#[[tool.hatch.envs.all.matrix]]
#python = ["3.11", "3.12"]
#
#[tool.hatch.envs.lint]
#detached = true
#dependencies = [
#  "black>=23.1.0",
#  "mypy>=1.0.0",
#  "ruff>=0.0.243",
#]
#[tool.hatch.envs.lint.scripts]
#typing = "mypy --install-types --non-interactive {args:spoofspy tests}"
#style = [
#  "ruff {args:.}",
#  "black --check --diff {args:.}",
#]
#fmt = [
#  "black {args:.}",
#  "ruff --fix {args:.}",
#  "style",
#]
#all = [
#  "style",
#  "typing",
#]
#
#[tool.black]
#target-version = ["py37"]
#line-length = 120
#skip-string-normalization = true
#
#[tool.ruff]
#target-version = "py37"
#line-length = 120
#select = [
#  "A",
#  "ARG",
#  "B",
#  "C",
#  "DTZ",
#  "E",
#  "EM",
#  "F",
#  "FBT",
#  "I",
#  "ICN",
#  "ISC",
#  "N",
#  "PLC",
#  "PLE",
#  "PLR",
#  "PLW",
#  "Q",
#  "RUF",
#  "S",
#  "T",
#  "TID",
#  "UP",
#  "W",
#  "YTT",
#]
#ignore = [
#  # Allow non-abstract empty methods in abstract base classes
#  "B027",
#  # Allow boolean positional values in function calls, like `dict.get(... True)`
#  "FBT003",
#  # Ignore checks for possible passwords
#  "S105", "S106", "S107",
#  # Ignore complexity
#  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
#]
#unfixable = [
#  # Don't touch unused imports
#  "F401",
#]
#
#[tool.ruff.isort]
#known-first-party = ["spoofspy"]
#
#[tool.ruff.flake8-tidy-imports]
#ban-relative-imports = "all"
#
#[tool.ruff.per-file-ignores]
## Tests can use magic values, assertions, and relative imports
#"tests/**/*" = ["PLR2004", "S101", "TID252"]
#
#[tool.coverage.run]
#source_pkgs = ["spoofspy", "tests"]
#branch = true
#parallel = true
#omit = [
#  "spoofspy/__about__.py",
#]
#
#[tool.coverage.paths]
#spoofspy = ["spoofspy", "*/spoofspy/spoofspy"]
#tests = ["tests", "*/spoofspy/tests"]
#
#[tool.coverage.report]
#exclude_lines = [
#  "no cov",
#  "if __name__ == .__main__.:",
#  "if TYPE_CHECKING:",
#]
